test(e2e): share the relay/preview process plumbing, skip the :3000 webServer for the collab project
- `tests/e2e/collab/process.ts` owns freePort / collectOutput /
  stopProcess; relay.ts and preview.ts import them instead of carrying
  verbatim copies.
- `playwright test --project=viewer-collab-e2e` alone no longer starts (or
  silently borrows from a sibling checkout) the shared :3000 webServer the
  project never uses; documented in the config and
  collaboration-testing.md.
- Spec header, viewer-page.ts and the evidence README say plainly that
  Copy is asserted enabled and the link read from the field (no headless
  clipboard), and that the private preview — not :3000 — is what is tested.
- Evidence refreshed from three consecutive green runs on the reviewed
  head: control 2 now takes 6 sockets (live + 5 backed-off probes) for a
  2.5 s frame hold instead of 12; acceptance numbers unchanged
  (184 / 313 / 313 / 1, texture 39bd94b5, export 111 nodes).
