+++ docs/architecture/evidence/share-seed-ready/result.json
@@ -1,6 +1,6 @@
 {
   "issue": 4446,
-  "startedAt": "2026-09-12T07:22:42.117Z",
+  "startedAt": "2026-09-12T07:55:54.715Z",
   "fixture": {
     "ifczipPath": "test-results/collab-share-seed/AC20-FZK-Haus.textured.ifczip",
     "wallExpressId": 15042,
@@ -13,34 +13,34 @@
     }
   },
   "relay": {
-    "wsUrl": "ws://127.0.0.1:52918",
+    "wsUrl": "ws://127.0.0.1:60880",
     "auth": "room-token (random COLLAB_TOKEN_SECRET)",
     "dataDir": "temp, deleted on stop"
   },
   "viewer": {
-    "url": "http://127.0.0.1:52923",
+    "url": "http://127.0.0.1:60886",
     "dist": "apps/viewer/dist (this checkout, private vite preview)"
   },
   "owner": {
-    "modelId": "370bc1f7-c6a0-4c78-ae33-5bd278511031",
+    "modelId": "228e96be-f54a-41a2-a5f0-b75431dbd8ab",
     "meshesLoaded": 317,
     "firstStatus": "Connecting to the room…",
     "seedPhases": [
       {
         "phase": "none",
-        "ms": 465
+        "ms": 520
       },
       {
         "phase": "syncing",
-        "ms": 136
+        "ms": 64
       },
       {
         "phase": "structure",
-        "ms": 36
+        "ms": 34
       },
       {
         "phase": "geometry",
-        "ms": 345
+        "ms": 333
       },
       {
         "phase": "confirming",
@@ -51,7 +51,7 @@
         "ms": null
       }
     ],
-    "copyEnabledAfterStatusMs": 837,
+    "copyEnabledAfterStatusMs": 834,
     "roomCounts": {
       "entities": 184,
       "geometryRecords": 313,
@@ -70,10 +70,10 @@
         "rgbaFnv1a": "39bd94b5"
       }
     },
-    "link": "http://127.0.0.1:52923/?room=289b4b402688&t=<redacted>",
-    "contextClosedAt": "2026-09-12T07:22:47.928Z",
+    "link": "http://127.0.0.1:60886/?room=39462317f6e6&t=<redacted>",
+    "contextClosedAt": "2026-09-12T07:56:00.561Z",
     "log": [
-      "http 404: http://127.0.0.1:52923/_vercel/insights/script.js",
+      "http 404: http://127.0.0.1:60886/_vercel/insights/script.js",
       "error: Failed to load resource: the server responded with a status of 404 (Not Found)",
       "log: [collab-ws] close code=1005 reason=\"\" clean=true"
     ]
@@ -117,7 +117,7 @@
       "summary": "Exported IFCX: 111 nodes, 90 meshes, 47 properties"
     },
     "log": [
-      "http 404: http://127.0.0.1:52923/_vercel/insights/script.js",
+      "http 404: http://127.0.0.1:60886/_vercel/insights/script.js",
       "error: Failed to load resource: the server responded with a status of 404 (Not Found)"
     ]
   },
@@ -155,7 +155,7 @@
       "provider": "indexeddb+websocket"
     },
     "log": [
-      "http 404: http://127.0.0.1:52923/_vercel/insights/script.js",
+      "http 404: http://127.0.0.1:60886/_vercel/insights/script.js",
       "error: Failed to load resource: the server responded with a status of 404 (Not Found)"
     ]
   },
@@ -202,7 +202,7 @@
       },
       "hydratedAtFirstCheck": 0,
       "log": [
-        "http 404: http://127.0.0.1:52923/_vercel/insights/script.js",
+        "http 404: http://127.0.0.1:60886/_vercel/insights/script.js",
         "error: Failed to load resource: the server responded with a status of 404 (Not Found)"
       ]
     },
@@ -216,28 +216,28 @@
   },
   "controlDelayedRelay": {
     "holdMs": 2500,
-    "heldFrames": 129,
-    "sockets": 12,
+    "heldFrames": 128,
+    "sockets": 6,
     "seedPhases": [
       {
         "phase": "none",
-        "ms": 182
+        "ms": 203
       },
       {
         "phase": "syncing",
-        "ms": 2599
+        "ms": 2580
       },
       {
         "phase": "structure",
-        "ms": 53
+        "ms": 44
       },
       {
         "phase": "geometry",
-        "ms": 348
+        "ms": 324
       },
       {
         "phase": "confirming",
-        "ms": 2687
+        "ms": 2834
       },
       {
         "phase": "ready",
@@ -287,8 +287,8 @@
     }
   },
   "relayLog": [
-    "[collab-server] listening at ws://127.0.0.1:52918 (data: C:\\Users\\LOUIST~1\\AppData\\Local\\Temp\\ifc-lite-relay-lYhDx3, auth: room-token, registry: off, blob-gc: every 360m)",
+    "[collab-server] listening at ws://127.0.0.1:60880 (data: C:\\Users\\LOUIST~1\\AppData\\Local\\Temp\\ifc-lite-relay-ZVxISx, auth: room-token, registry: off, blob-gc: every 360m)",
     "[blob-gc] swept 0 room log(s): deleted 0 blob(s), 0 stale upload(s), kept 0"
   ],
-  "finishedAt": "2026-09-12T07:23:16.430Z"
+  "finishedAt": "2026-09-12T07:56:28.708Z"
 }
