+++ tests/e2e/collab-share-seed.e2e.spec.ts
@@ -11,7 +11,9 @@
  * `@ifc-lite/collab-server` (`collab/relay.ts`), both on ephemeral ports. The
  * owner loads the textured AC20-FZK-Haus IFCZIP (`collab/textured-ac20.ts`
  * builds it from the plain fixture), opens File → Share, and its context is
- * closed the instant the Copy action is enabled — no grace period. A fresh
+ * closed the instant the Copy action is enabled — no grace period. Copy is
+ * asserted enabled and the link is read from the field; the clipboard itself
+ * is not exercised headless. A fresh
  * guest must then reconstruct the complete room (every entity, every geometry
  * reference, the textured wall byte-identical), a third context must see the
  * same on rejoin, and the guest's ordinary IFC export must produce a download.
